Translated text file system to JSON, updated reality bot and readme.
