+++ README.md
@@ -1,9 +1,8 @@
-# Python-Bot-
+# Python-Bot
 
 This cheeky lil boi counts the number of times our teacher says the statement "So the reality is"
 - You can add another reality by typing into discord "?reality", and an extra reality will be added to the reality count
-- You can call back on how many realities there are with "?mh count", and the current number of realities will be displayed
-- The dates that an extra reality the was created can be called back upon on when with "?date + [date you want to find]", and the number of realities that were added on that day will be added.
+- You can call back on how many realities there are with "?r count", and the current number of realities will be displayed
+- The dates that an extra reality the was created can be called back upon on when with "?r date + [date you want to find]", and the number of realities that were added on that day will be added.
 
-The file counter.txt holds the value for the reality count so that it is not lost each time to program stops running.
-The file date.txt holds that data that eah reality was created.
+The file counter.json holds the value for the reality count per each day so that it is not lost each time to program stops running.
